+++ src/tasks/job.ts
@@ -1,26 +1,10 @@
-interface JobConstructorOpts {
+export interface JobConstructorOpts {
   id: number
   source: string
   lang: string
   scenario: "run" | "submit" | "project"
   timelimit?: number
 } 
-interface RunJobConstructorOpts extends JobConstructorOpts {
-  stdin: string
-}
-interface TestcaseOpts {
-  id: number, 
-  input: string, 
-  output: string
-}
-interface SubmitJobConstructorOpts extends JobConstructorOpts {
-  testcases: Array<TestcaseOpts>
-}
-
-interface ProjectConstructorOpts extends JobConstructorOpts {
-  problem: string,
-  submissionDirs: string
-}
 
 export class Job { 
   id: number
@@ -38,31 +22,4 @@ export class Job {
   }
 }
 
-export class RunJob extends Job {
-  stdin: string
 
-  constructor({ id, source, lang, timelimit, scenario, stdin }: RunJobConstructorOpts) {
-    super({id, source, lang, timelimit, scenario})
-    this.stdin = stdin
-  }
-}
-
-export class SubmitJob extends Job {
-  testcases: Array<TestcaseOpts>
-
-  constructor({ id, source, lang, timelimit, scenario, testcases }: SubmitJobConstructorOpts) {
-    super({id, source, lang, timelimit, scenario})
-    this.testcases = testcases
-  }
-}
-
-export class ProjectJob extends Job {
-  problem: string
-  submissionDirs: string
-
-  constructor({ id, source, lang, timelimit, scenario, problem, submissionDirs}: ProjectConstructorOpts) {
-    super({ id, source, lang, timelimit, scenario})
-    this.problem = problem
-    this.submissionDirs = submissionDirs
-  }
-}
